Added "etr_factor" parameter for scaling the ETr values (mainly GRIDMET)
I'm not sure exactly where to apply this in the interpolation function.  To avoid having a separate map call for this, I added it into the aggregations functions.  The problem with this approach is that it would be very difficult to apply different factors if a composite ETr source was used (but this isn't supported right now).